working delete avatar

+++ main.py
@@ -301,7 +301,7 @@ def get_image(user_id):
             return RESPONSE_403, 403
         user_key = client.key(USERS, user_id)
         user = client.get(key=user_key)
-        if 'avatar_file' not in user.keys():
+        if 'avatar_file' not in user.keys() or user['avatar_file'] == "":
             return RESPONSE_404, 404
         file_name = user['avatar_file']
         storage_client = storage.Client()
@@ -316,14 +316,31 @@ def get_image(user_id):
         return RESPONSE_401, 401
 
 
-@app.route('/images/<file_name>', methods=['DELETE'])
-def delete_image(file_name):
-    storage_client = storage.Client()
-    bucket = storage_client.get_bucket(AVATAR_BUCKET)
-    blob = bucket.blob(file_name)
-    # Delete the file from Cloud Storage
-    blob.delete()
-    return '', 204
+@app.route('/' + USERS + '/<int:user_id>/avatar', methods=['DELETE'])
+def delete_image(user_id):
+    try:
+        payload = verify_jwt(request)
+        sub = payload["sub"]
+        user_based_on_jwt = get_user(sub)
+        if user_based_on_jwt['id'] != user_id:
+            return RESPONSE_403, 403
+        user_key = client.key(USERS, user_id)
+        user = client.get(key=user_key)
+        if 'avatar_file' not in user.keys() or user['avatar_file'] == "":
+            return RESPONSE_404, 404
+        file_name = user['avatar_file']
+        storage_client = storage.Client()
+        bucket = storage_client.get_bucket(AVATAR_BUCKET)
+        blob = bucket.blob(file_name)
+        # Delete the file from Cloud Storage
+        blob.delete()
+        user.update({
+            'avatar_file': ""
+        })
+        client.put(user)
+        return '', 204
+    except AuthError:
+        return RESPONSE_401, 401
 
 
 if __name__ == '__main__':
